refactor: Move constants to separate constants file
