+++ src/services/data.ts
@@ -1,18 +1,5 @@
-import { ExtensionData, Settings } from '../types';
-
-export const defaultSettings: Settings = {
-  pomodoro: {
-    workTime: 25,
-    breakTime: 5,
-  },
-};
-
-export const initialData: ExtensionData = {
-  settings: {
-    ...defaultSettings,
-  },
-  notes: [],
-};
+import { ExtensionData } from '../types';
+import { initialData } from '../constants';
 
 function isValidKey(
   key: string | string[]
